YAY-5 fix: change name,  transfer schemas, change onQueryStarted
